refactor(core): relocate lbfgs_minimize to variational/lbfgs.rs (#74 A1)
Pure relocation, byte-identical math. Moves lbfgs_minimize (and its private
`dot` helper, used only by it) out of dmr.rs into the new variational/lbfgs.rs,
re-exported as crate::variational::lbfgs_minimize. Re-points the 6 importers
(ctm, sts, dtm, etm, sage, and dmr itself). No fit algorithm changed.

cargo test --lib: 60 passed (the ctm/sts determinism + finite-difference tests
exercise lbfgs); compiles under --features python.
